refactor: extract runtime execution loop
