+++ lensnode/lensnode/agent_runtime/execution.py
@@ -1,0 +1,536 @@
+"""Agent streaming, wrap-up synthesis, and event emission."""
+
+import logging
+
+from langchain_core.messages import HumanMessage
+
+from ..agent_tools import SELF_REPORTING_TOOLS
+from ..gateway_model import RunCancelledError
+from .messages import (
+    extract_final_message as _extract_final_message,
+    normalize_plan_steps as _normalize_plan_steps,
+    tool_call_summary as _tool_call_summary,
+)
+from .prompts import (
+    answer_language_requirement as _answer_language_requirement,
+    pick_text as _pick_text,
+)
+
+LOGGER = logging.getLogger("lensnode")
+class EmptyAgentResponseError(RuntimeError):
+    """The agent and its one recovery attempt both returned no text."""
+
+    code = "EMPTY_AGENT_RESPONSE"
+
+def _run_agent_with_turn_limit(
+    agent,
+    messages,
+    max_turns,
+    model=None,
+    thread=None,
+    turn_baseline_ai=None,
+    event_baseline_ai=None,
+    resume_from_checkpoint=False,
+    emit_event=None,
+    answer_language="English",
+    cancel_event=None,
+    wrapup_event=None,
+    token_budget_wrapup_event=None,
+    on_checkpoint_state=None,
+    input_checkpoint_seeded=False,
+):
+    """Stream agent events and stop after max_turns NEW AI turns.
+
+    `messages` may be prefixed with prior conversation turns. Historical
+    assistant turns are excluded from both the turn count and event
+    emission, so the limit and trace reflect only the current run.
+
+    On resume, ``turn_baseline_ai`` remains the original history count so
+    turns already consumed by this Run still count. ``event_baseline_ai``
+    suppresses re-emitting model events already present in the checkpoint.
+
+    A subagent ``task`` delegation runs to completion inside a single
+    parent-graph step (deepagents calls it via a plain, synchronous
+    ``subagent.invoke(...)``), so the turn count can only be checked
+    between parent steps — a delegation in flight when the budget is
+    reached can overshoot max_turns by however many turns it used
+    internally before returning. This is expected, not a bug in the
+    count itself; it is the reason a forced wrap-up (see
+    _synthesize_wrapup_answer) matters more than counting precisely.
+
+    Returns (answer, truncated, termination_reason). ``truncated`` is true
+    when the agent was stopped before it finished naturally, and the reason
+    preserves the gate that requested wrap-up.
+    """
+
+    last_state = {"messages": messages} if resume_from_checkpoint else None
+    truncated = False
+    truncation_reason = None
+    seen_tool_calls = set()
+    seen_model_calls = set()
+    plan_state = {"revision": 0}
+    baseline_ai = (
+        turn_baseline_ai
+        if turn_baseline_ai is not None
+        else sum(1 for m in messages if m.get("role") == "assistant")
+    )
+    emitted_baseline_ai = (
+        event_baseline_ai
+        if event_baseline_ai is not None
+        else baseline_ai
+    )
+    seeded_baseline = False
+
+    if resume_from_checkpoint:
+        graph_input = None
+    elif input_checkpoint_seeded:
+        graph_input = {"messages": []}
+    else:
+        graph_input = {"messages": messages}
+    for state in agent.stream(
+        graph_input,
+        stream_mode="values",
+        config={
+            "recursion_limit": 500,
+            **(thread or {}),
+        },
+    ):
+        if cancel_event is not None and cancel_event.is_set():
+            raise RunCancelledError(
+                "Run was cancelled; stopping the agent loop."
+            )
+        last_state = state
+        if on_checkpoint_state is not None:
+            on_checkpoint_state()
+        current = state.get("messages", [])
+        if not seeded_baseline:
+            # Seed the historical assistant turns by their (now-assigned)
+            # message and tool-call ids so they are never emitted or counted
+            # as new turns. Reusing the tool event reducer also restores its
+            # plan revision/state without publishing the old events again.
+            baseline_messages = []
+            ai_count = 0
+            for message in current:
+                if getattr(message, "type", "") != "ai":
+                    continue
+                ai_count += 1
+                if ai_count <= emitted_baseline_ai:
+                    baseline_messages.append(message)
+                    seen_model_calls.add(
+                        getattr(message, "id", None) or id(message)
+                    )
+            _emit_new_tool_calls(
+                baseline_messages,
+                seen_tool_calls,
+                lambda _name, _detail: None,
+                plan_state=plan_state,
+            )
+            seeded_baseline = True
+        if emit_event is not None:
+            _emit_new_model_calls(current, seen_model_calls, emit_event)
+            _emit_new_tool_calls(
+                current,
+                seen_tool_calls,
+                emit_event,
+                plan_state=plan_state,
+            )
+        ai_turns = sum(
+            1
+            for m in current
+            if getattr(m, "type", "") == "ai"
+        ) - baseline_ai
+        if wrapup_event is not None and wrapup_event.is_set():
+            truncated = True
+            truncation_reason = "soft_deadline"
+            if emit_event is not None:
+                emit_event("deepagents.agent.soft_deadline", {})
+            break
+        if (
+            token_budget_wrapup_event is not None
+            and token_budget_wrapup_event.is_set()
+        ):
+            truncated = True
+            truncation_reason = "token_budget_wrapup"
+            if emit_event is not None:
+                emit_event("deepagents.agent.token_budget", {})
+            break
+        if ai_turns >= max_turns:
+            truncated = True
+            truncation_reason = "turn_limit"
+            break
+
+    answer = _extract_final_message(last_state or {})
+    force_wrapup = truncation_reason in {
+        "soft_deadline",
+        "token_budget_wrapup",
+    }
+    needs_wrapup = force_wrapup or not answer.strip()
+    if truncated and model is not None and needs_wrapup:
+        # The cutoff landed mid-turn (e.g. right after a tool call, before
+        # any answer text), so there is nothing to extract — but the
+        # conversation likely still holds real findings from every prior
+        # turn. Ask once more, without tools, for a best-effort synthesis
+        # instead of discarding all of that work.
+        synthesis = _synthesize_wrapup_answer(
+            model,
+            (last_state or {}).get("messages", []),
+            answer_language,
+            emit_event,
+            reason=truncation_reason or "limit",
+        )
+        if synthesis:
+            answer = synthesis
+    if not truncated and not answer.strip() and model is not None:
+        answer = _synthesize_wrapup_answer(
+            model,
+            (last_state or {}).get("messages", []),
+            answer_language,
+            emit_event,
+            reason="empty",
+        )
+    if not answer.strip():
+        raise EmptyAgentResponseError(
+            "Agent returned no answer after one recovery attempt."
+        )
+    if truncated and answer.strip():
+        if truncation_reason == "soft_deadline":
+            answer += _pick_text(
+                "\n\n---\n*即将达到硬截止时间，以上回答由当前已有证据"
+                "综合生成，调查可能尚未完全完成。*",
+                "\n\n---\n*Approaching the hard deadline, this answer was "
+                "synthesized from the evidence already collected and the "
+                "investigation may be incomplete.*",
+                answer_language,
+            )
+        else:
+            answer += _pick_text(
+                "\n\n---\n*已达到当前分析深度上限，本次调查未完全完成。"
+                "如需更完整的结果，请调高分析档位后重试。*",
+                "\n\n---\n*Reached the current analysis-depth limit before "
+                "the investigation fully completed. Raise the analysis "
+                "tier for a more complete result.*",
+                answer_language,
+            )
+    termination_reason = truncation_reason
+    if termination_reason is None and model is not None:
+        model_reason = getattr(model, "stop_reason", None)
+        if model_reason in {
+            "loop_capped",
+            "model_length_capped",
+            "safety_terminated",
+            "token_budget_wrapup",
+            "token_capped",
+        }:
+            termination_reason = model_reason
+    return answer, truncated, termination_reason
+
+
+def _strip_dangling_tool_call(messages):
+    """Drop a trailing AI message with unresolved tool_calls.
+
+    A truncated loop can stop right after the model requested a tool call
+    but before the tool result was recorded. Most providers reject a
+    follow-up call whose history ends on a tool_calls message with no
+    matching tool response, so that dangling turn is dropped before
+    asking for a wrap-up answer — every earlier, resolved turn is kept.
+    """
+
+    if not messages:
+        return messages
+    last = messages[-1]
+    if getattr(last, "type", "") == "ai" and getattr(
+        last, "tool_calls", None
+    ):
+        return messages[:-1]
+    return messages
+
+
+def _synthesize_wrapup_answer(
+    model,
+    current,
+    answer_language,
+    emit_event,
+    reason="limit",
+):
+    """Ask once for a tool-free answer after cutoff or an empty terminal.
+
+    The prompt asks the model to synthesize its best answer from the current
+    conversation, so prior work is not discarded. Returns "" if this call
+    fails; the caller decides whether existing partial text is sufficient or
+    an explicit empty-response error should be raised.
+
+    A RunCancelledError is deliberately NOT caught here: model.invoke()
+    checks cancel_event at the top of the gateway call
+    (LensGatewayChatModel._check_cancelled), so a cancellation landing in
+    the narrow window between the turn-limit loop exiting and this call
+    starting must still stop the run, not be swallowed into an empty
+    "wrap-up failed" result.
+    """
+
+    if reason == "empty":
+        instruction = _pick_text(
+            "你上一轮没有输出可见答案。不要调用任何工具，请仅基于当前对话"
+            "和已取得的结果，直接向用户给出完整答案。",
+            "Your previous turn produced no visible answer. Do not call "
+            "tools. Based only on the current conversation and collected "
+            "results, write the complete answer to the user now.",
+            answer_language,
+        )
+    elif reason == "token_budget_wrapup":
+        instruction = _pick_text(
+            "你已经达到本次调查的 Token 预算，不能再调用任何工具。请仅基于"
+            "当前对话和已取得的结果，直接给出最完整的最终答案；未确认或未覆盖"
+            "的部分必须明确说明。不要向用户提及内部 Token 预算或运行控制状态。",
+            "You have reached the token budget for this investigation and "
+            "cannot call more tools. Based only on the current conversation "
+            "and collected results, write the most complete final answer. "
+            "Clearly identify anything unconfirmed or not covered. Do not "
+            "mention internal token budgets or runtime control state.",
+            answer_language,
+        )
+    elif reason == "execution_failed":
+        instruction = _pick_text(
+            "能力执行已达到恢复上限，不能再调用任何工具。请仅基于当前"
+            "对话和已取得的信息给出简洁回答，并明确说明执行失败及未完成"
+            "的部分。",
+            "Capability execution reached its recovery limit. Do not call "
+            "tools. Give a concise answer from the current conversation and "
+            "collected information, and state what failed or remains undone.",
+            answer_language,
+        )
+    else:
+        instruction = _pick_text(
+            "你已经达到本次分析的步数上限，不能再调用任何工具。请基于目前"
+            "为止已经掌握的全部信息，直接给出你能给出的最完整回答。如果"
+            "调查还有尚未确认或未覆盖到的部分，请明确说明。",
+            "You have reached the step limit for this analysis and cannot "
+            "call any more tools. Based on everything you have gathered so "
+            "far, write the most complete answer you can now. Clearly note "
+            "any part of the investigation you were not able to confirm.",
+            answer_language,
+        )
+    instruction = (
+        f"{instruction}\n\n"
+        f"{_answer_language_requirement(answer_language)}"
+    )
+    wrapup_messages = _strip_dangling_tool_call(current) + [
+        HumanMessage(content=instruction)
+    ]
+    if emit_event is not None:
+        event = (
+            "deepagents.answer.recovery"
+            if reason == "empty"
+            else "deepagents.agent.wrapup"
+        )
+        emit_event(event, {})
+    try:
+        response = model.invoke(
+            wrapup_messages,
+            runtime_final_synthesis=True,
+        )
+    except RunCancelledError:
+        raise
+    except Exception:
+        LOGGER.exception("Wrap-up synthesis call failed after truncation")
+        return ""
+    content = getattr(response, "content", None)
+    if isinstance(content, str):
+        return content.strip()
+    return str(content or "").strip()
+
+
+def _model_summary(message, limit=160):
+    """Return a short preview of a model turn for the trace.
+
+    Prefers the assistant's own text; when the turn only issued tool
+    calls (no text), shows the tools it decided to call instead.
+    """
+
+    content = getattr(message, "content", "")
+    if isinstance(content, list):
+        parts = []
+        for block in content:
+            if isinstance(block, dict):
+                parts.append(str(block.get("text") or ""))
+            else:
+                parts.append(str(block))
+        content = " ".join(parts)
+    text = " ".join(str(content or "").split())
+    if text:
+        return text[:limit] + ("…" if len(text) > limit else "")
+    calls = [
+        call.get("name")
+        for call in (getattr(message, "tool_calls", None) or [])
+        if call.get("name")
+    ]
+    if calls:
+        ordered, counts = [], {}
+        for name in calls:
+            if name not in counts:
+                ordered.append(name)
+            counts[name] = counts.get(name, 0) + 1
+        parts = [
+            f"{name}×{counts[name]}" if counts[name] > 1 else name
+            for name in ordered
+        ]
+        return "→ " + ", ".join(parts)
+    return ""
+
+
+def _emit_new_model_calls(messages, seen, emit_event):
+    """Emit an event for each new AI response (one LLM round).
+
+    Each AI message is one round-trip to the model. The gateway returns
+    token usage in the message's response_metadata, so surfacing these
+    makes every LLM call visible in the trace and attributes token usage
+    to the run. Dedup keys on the stable message id rather than position,
+    so summarization (which rewrites the message list) cannot make a new
+    final turn collide with an already-seen positional index.
+    """
+
+    for message in messages:
+        if getattr(message, "type", "") != "ai":
+            continue
+        key = getattr(message, "id", None) or id(message)
+        if key in seen:
+            continue
+        seen.add(key)
+        meta = getattr(message, "response_metadata", None) or {}
+        usage = meta.get("usage") or {}
+        emit_event(
+            "llm.response",
+            {
+                "round": len(seen),
+                "model": usage.get("model"),
+                "prompt_tokens": usage.get("prompt_tokens"),
+                "completion_tokens": usage.get("completion_tokens"),
+                "total_tokens": usage.get("total_tokens"),
+                "cached_tokens": (
+                    usage.get("cached_tokens")
+                    or usage.get("prompt_cache_hit_tokens")
+                    or 0
+                ),
+                "reasoning_tokens": usage.get("reasoning_tokens") or 0,
+                "cost": usage.get("cost"),
+                "latency_ms": meta.get("latency_ms"),
+                "finish_reason": meta.get("finish_reason"),
+                "stop_reason": _response_stop_reason(meta),
+                "run_token_usage": meta.get("run_token_usage"),
+                "summary": _model_summary(message),
+            },
+        )
+
+
+def _response_stop_reason(metadata):
+    """Return the normalized runtime stop reason for one model response."""
+
+    for key, reason in (
+        ("token_capped", "token_capped"),
+        ("loop_capped", "loop_capped"),
+        ("safety_terminated", "safety_terminated"),
+        ("model_length_capped", "model_length_capped"),
+    ):
+        if metadata.get(key):
+            return reason
+    return None
+
+
+def _emit_new_tool_calls(
+    messages,
+    seen,
+    emit_event,
+    *,
+    plan_state=None,
+):
+    """Emit a progress event for each not-yet-seen agent tool call.
+
+    The built-in workspace tools emit their own start/done events, but
+    the Deep Agent loop also calls model-driven tools (write_file, ls,
+    task delegation, MCP tools) that are otherwise invisible. Surfacing
+    every tool call here lets the frontend show real progress instead of
+    a frozen status during long turns.
+    """
+
+    for message in messages:
+        if getattr(message, "type", "") != "ai":
+            continue
+        for call in getattr(message, "tool_calls", None) or []:
+            call_id = call.get("id") or ""
+            if not call_id or call_id in seen:
+                continue
+            seen.add(call_id)
+            name = call.get("name") or "tool"
+            if name == "write_todos":
+                state = plan_state if plan_state is not None else {
+                    "revision": 0
+                }
+                state["revision"] = int(state.get("revision") or 0) + 1
+                todos = (call.get("args") or {}).get("todos") or []
+                incoming_steps = _normalize_plan_steps(todos)
+                initial_steps = state.get("steps") or []
+                if initial_steps:
+                    incoming_by_id = {
+                        item["id"]: item for item in incoming_steps
+                    }
+                    steps = [
+                        {
+                            **item,
+                            "status": incoming_by_id.get(
+                                item["id"], item
+                            )["status"],
+                        }
+                        for item in initial_steps
+                    ]
+                else:
+                    steps = incoming_steps
+                state["steps"] = steps
+                emit_event(
+                    "workflow.plan.updated",
+                    {
+                        "event_type": "plan.updated",
+                        "visibility": "user",
+                        "payload": {
+                            "revision": state["revision"],
+                            "steps": steps,
+                        },
+                    },
+                )
+                if (
+                    todos
+                    and all(
+                        item.get("status") == "completed" for item in todos
+                    )
+                    and not state.get("answering_phase_emitted")
+                ):
+                    state["answering_phase_emitted"] = True
+                    state["execution_phase_emitted"] = True
+                    emit_event(
+                        "workflow.phase.changed",
+                        {
+                            "event_type": "phase.changed",
+                            "visibility": "user",
+                            "payload": {"phase": "answering"},
+                        },
+                    )
+                continue
+            if (
+                plan_state is not None
+                and not plan_state.get("execution_phase_emitted")
+            ):
+                plan_state["execution_phase_emitted"] = True
+                emit_event(
+                    "workflow.phase.changed",
+                    {
+                        "event_type": "phase.changed",
+                        "visibility": "user",
+                        "payload": {"phase": "executing"},
+                    },
+                )
+            if name in SELF_REPORTING_TOOLS:
+                # avoids a duplicate trace line; the tool emits its own
+                # .start/.done with a richer argument summary
+                continue
+            emit_event(
+                f"tool.{name}.invoke",
+                {"tool": name, "summary": _tool_call_summary(call)},
+            )
